# docs/api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
...
  "paths" : {
    "/api/prof/cours/{coursId}" : {
      "get" : {
        "tags" : [ "Édition de cours" ],
        "summary" : "Charge un cours pour l'édition",
        "description" : "Renvoie le brouillon en cours (métadonnées, contenu, classes portées). Réservé à l'auteur du cours.",
        "operationId" : "charger",
        "parameters" : [ {
          "name" : "coursId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Cours chargé",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CoursDetailResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Le cours ne vous appartient pas",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CoursDetailResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "Cours introuvable",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CoursDetailResponse"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Édition de cours" ],
        "summary" : "Enregistre le brouillon (titre + contenu)",
        "operationId" : "enregistrer",
        "parameters" : [ {
          "name" : "coursId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ModifierCoursRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "Brouillon enregistré"
          },
          "403" : {
            "description" : "Le cours ne vous appartient pas"
          },
          "404" : {
            "description" : "Cours introuvable"
          }
        }
      }
    },
    "/api/prof/cours/{coursId}/portees" : {
      "put" : {
        "tags" : [ "Édition de cours" ],
        "summary" : "Définit les classes auxquelles le cours est porté",
        "description" : "Chaque classe doit être une classe de l'enseignant connecté.",
        "operationId" : "definirPortees",
        "parameters" : [ {
          "name" : "coursId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DefinirPorteesRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "Portées enregistrées"
          },
          "403" : {
            "description" : "Cours non détenu, ou classe hors de vos classes"
          },
          "404" : {
            "description" : "Cours introuvable"
          }
        }
      }
    },
...
    "/api/prof/cours" : {
      "get" : {
        "tags" : [ "Édition de cours" ],
        "summary" : "Liste vos cours",
        "description" : "Les cours dont vous êtes l'auteur, du plus récemment modifié au plus ancien.",
        "operationId" : "mesCours",
        "responses" : {
          "200" : {
            "description" : "Vos cours",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/CoursResumeResponse"
                  }
                }
              }
            }
          },
          "401" : {
            "description" : "Authentification requise",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/CoursResumeResponse"
                  }
                }
              }
            }
          },
          "403" : {
            "description" : "Réservé aux enseignants",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/CoursResumeResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Édition de cours" ],
        "summary" : "Crée un brouillon de cours",
        "description" : "L'auteur est l'enseignant connecté, jamais lu dans la requête. L'établissement est le sien ; s'il en a plusieurs, la requête précise lequel. Le contenu est validé à la publication, pas ici.",
        "operationId" : "creer_1",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreerCoursRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Brouillon créé",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CoursCreeResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Authentification requise",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CoursCreeResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Réservé aux enseignants, dans un de leurs établissements",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CoursCreeResponse"
                }
              }
            }
          },
          "422" : {
            "description" : "Établissement indéterminé : aucun, ou plusieurs sans précision",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CoursCreeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/prof/cours/{coursId}/publier" : {
      "post" : {
        "tags" : [ "Édition de cours" ],
        "summary" : "Publie le cours",
        "description" : "Valide le brouillon et fige une version immuable visible des classes portées. Les réponses de correction sont retirées du contenu figé.",
        "operationId" : "publier",
        "parameters" : [ {
          "name" : "coursId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Cours publié ; version figée",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PublicationResult"
                }
              }
            }
          },
          "403" : {
            "description" : "Le cours ne vous appartient pas",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PublicationResult"
                }
              }
            }
          },
          "404" : {
            "description" : "Cours introuvable",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PublicationResult"
                }
              }
            }
          },
          "422" : {
            "description" : "Le brouillon ne respecte pas le schéma / le référentiel",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PublicationResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/cours/{coursId}" : {
      "get" : {
        "tags" : [ "Contenu" ],
        "summary" : "Récupère un cours d'enseignant publié",
        "description" : "Retourne la version publiée du cours, dans le même format qu'un chapitre de catalogue. Réservé aux élèves des classes portées. Les réponses correctes des exercices ne sont jamais incluses.",
        "operationId" : "getCoursPublie",
        "parameters" : [ {
          "name" : "coursId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Cours trouvé et accessible",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/JsonNode"
                }
              }
            }
          },
          "401" : {
            "description" : "Authentification requise",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/JsonNode"
                }
              }
            }
          },
          "403" : {
            "description" : "Cours non accessible à l'élève (hors portée)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/JsonNode"
                }
              }
            }
          },
          "404" : {
            "description" : "Cours introuvable ou non publié",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/JsonNode"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "JsonNode" : { },
      "ModifierCoursRequest" : {
        "type" : "object",
        "properties" : {
          "titre" : {
            "type" : "string",
            "maxLength" : 300,
            "minLength" : 0
          },
          "content" : {
            "$ref" : "#/components/schemas/JsonNode"
          }
        },
        "required" : [ "titre" ]
      },
      "DefinirPorteesRequest" : {
        "type" : "object",
        "properties" : {
          "classeIds" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            },
            "uniqueItems" : true
          }
        },
        "required" : [ "classeIds" ]
      },
...
      "CreerCoursRequest" : {
        "type" : "object",
        "properties" : {
          "titre" : {
            "type" : "string",
            "maxLength" : 300,
            "minLength" : 0
          },
          "niveauCode" : {
            "type" : "string",
            "maxLength" : 16,
            "minLength" : 0
          },
          "matiereCode" : {
            "type" : "string",
            "maxLength" : 64,
            "minLength" : 0
          },
          "etablissementId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "content" : {
            "$ref" : "#/components/schemas/JsonNode"
          }
        },
        "required" : [ "matiereCode", "niveauCode", "titre" ]
      },
      "CoursCreeResponse" : {
        "type" : "object",
        "properties" : {
          "coursId" : {
            "type" : "string",
            "format" : "uuid"
          }
        }
      },
      "PublicationResult" : {
        "type" : "object",
        "properties" : {
          "coursId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "version" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "CoursResumeResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "titre" : {
            "type" : "string"
          },
          "niveauCode" : {
            "type" : "string"
          },
          "matiereCode" : {
            "type" : "string"
          },
          "statut" : {
            "type" : "string"
          },
          "versionPubliee" : {
            "type" : "integer",
            "format" : "int32"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "CoursDetailResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "titre" : {
            "type" : "string"
          },
          "niveauCode" : {
            "type" : "string"
          },
          "matiereCode" : {
            "type" : "string"
          },
          "statut" : {
            "type" : "string"
          },
          "versionPubliee" : {
            "type" : "integer",
            "format" : "int32"
          },
          "content" : {
            "$ref" : "#/components/schemas/JsonNode"
          },
          "classeIds" : {
            "type" : "array",
            "items" : {
              "type" : "string",
              "format" : "uuid"
            },
            "uniqueItems" : true
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
    }
  }
}